mqtt_webrtc_test/requirements.txt
numpy==2.2.6
paho-mqtt
pymavlink
aiortc
mosquitto
mosquitto-clients